+++ src/main/java/ba/ibu/edu/bemytech/core/model/OrderedProduct.java
@@ -1,0 +1,23 @@
+package ba.ibu.edu.bemytech.core.model;
+
+public class OrderedProduct {
+    private Product product;
+
+    private int quantity;
+
+    public Product getProduct() {
+        return product;
+    }
+
+    public void setProduct(Product product) {
+        this.product = product;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+}
